+++ app/src/test/java/com/codesoom/assignment/user/domain/UserTest.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.domain.user;
+package com.codesoom.assignment.user.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
